+++ src/main/java/com/sparta/blushmarket/dto/LoginRequestDto.java
@@ -4,6 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class LoginRequestDto {
-    private String name;
+    private String email;
     private String password;
 }
