+++ src/main/java/com/sparta/blushmarket/dto/PostResponseDto.java
@@ -1,42 +1,42 @@
 package com.sparta.blushmarket.dto;
 
-import com.sparta.blushmarket.entity.Likes;
 import com.sparta.blushmarket.entity.Post;
-import com.sparta.blushmarket.entity.SellState;
+import com.sparta.blushmarket.entity.enumclass.SellState;
 import lombok.Builder;
 import lombok.Getter;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Getter
 public class PostResponseDto {
     private Long id;
+    private String username;
     private String title;
     private String content;
     private String image;
     private SellState sellState;
 
     private boolean likes;
-    private List<CommentResponseDto> commentList;
+
 
 
     @Builder
-    private PostResponseDto(boolean likes, Post post, List<CommentResponseDto> commentList) {
+    private PostResponseDto(boolean likes, Post post) {
         this.id = post.getId();
+        this.username =post.getMember().getName();
         this.title = post.getTitle();
         this.content = post.getContent();
         this.image = post.getImage();
         this.sellState = post.getSellState();
         this.likes = likes;
-        this.commentList = commentList;
+
     }
 
-    public static PostResponseDto from(boolean likes,Post post,List<CommentResponseDto>  commentList) {
+    public static PostResponseDto from(boolean likes, Post post) {
         return PostResponseDto.builder()
                 .post(post)
                 .likes(likes)
-                .commentList(commentList)
+
                 .build();
 
     }
