+++ src/main/java/com/sparta/blushmarket/entity/enumclass/SellState.java
@@ -1,4 +1,4 @@
-package com.sparta.blushmarket.entity;
+package com.sparta.blushmarket.entity.enumclass;
 
 import com.sparta.blushmarket.exception.CustomException;
